# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List published opportunities
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/OpportunityCategory"
      responses:
        "200":
          description: Matching opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get a published opportunity
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404": { $ref: "#/components/responses/NotFound" }
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List published programs
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Matching programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a published program
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404": { $ref: "#/components/responses/NotFound" }
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List published events
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Matching events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a published event
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404": { $ref: "#/components/responses/NotFound" }
...
  /contact-messages:
    post:
      operationId: sendContactMessage
      tags: [contact]
      summary: Send a message to B4P CODEFOUND
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessageReceipt"
...
  /admin/updates:
    get:
      operationId: listAdminUpdates
      tags: [admin, updates]
      summary: List updates, including drafts
      responses:
        "200":
          description: Updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Update"
    post:
      operationId: createUpdate
      tags: [admin, updates]
      summary: Create and optionally publish an update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInput"
      responses:
        "201":
          description: Created update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Update"
  /admin/updates/{id}:
    patch:
      operationId: updateUpdate
      tags: [admin, updates]
      summary: Edit, publish or unpublish an update
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUpdate"
      responses:
        "200":
          description: Updated update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Update"
    delete:
      operationId: deleteUpdate
      tags: [admin, updates]
      summary: Delete an update
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Update deleted }
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    NotFound:
      description: Record not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    OpportunityCategory:
      type: string
      enum: [Jobs, Internships, Scholarships, Fellowships, Trainings, Grants, Volunteer, Other]
...
    UpdateCategory:
      type: string
      enum: [Announcement, News, Important notice, Program update, Opportunity update, Event reminder]
...
    Opportunity:
      type: object
      required: [id, title, category, organization, contactInfo, coverImage, description, requirements, location, deadline, applicationUrl, datePosted, published]
      properties:
        id: { type: integer }
        title: { type: string }
        category: { $ref: "#/components/schemas/OpportunityCategory" }
        organization: { type: string }
        contactInfo: { type: ["string", "null"] }
        coverImage: { type: ["string", "null"], format: uri }
        description: { type: string }
        requirements: { type: string }
        location: { type: string }
        deadline: { type: ["string", "null"], format: date }
        applicationUrl: { type: ["string", "null"], format: uri }
        datePosted: { type: string, format: date-time }
        published: { type: boolean }
...
    Program:
      type: object
      required: [id, name, coverImage, description, objectives, eligibility, location, startDate, endDate, applicationDeadline, published]
      properties:
        id: { type: integer }
        name: { type: string }
        coverImage: { type: ["string", "null"], format: uri }
        description: { type: string }
        objectives: { type: string }
        eligibility: { type: string }
        location: { type: string }
        startDate: { type: ["string", "null"], format: date }
        endDate: { type: ["string", "null"], format: date }
        applicationDeadline: { type: ["string", "null"], format: date }
        published: { type: boolean }
...
    Event:
      type: object
      required: [id, title, imageUrl, date, registrationDeadline, registrationUrl, time, location, description, published]
      properties:
        id: { type: integer }
        title: { type: string }
        imageUrl: { type: ["string", "null"], format: uri }
        date: { type: string, format: date }
        registrationDeadline: { type: ["string", "null"], format: date }
        registrationUrl: { type: ["string", "null"], format: uri }
        time: { type: string }
        location: { type: string }
        description: { type: string }
        published: { type: boolean }
...
    Update:
      type: object
      required: [id, title, category, content, imageUrl, published, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        category: { $ref: "#/components/schemas/UpdateCategory" }
        content: { type: string }
        imageUrl: { type: ["string", "null"], format: uri }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UpdateInput:
      type: object
      required: [title, category, content, imageUrl, published]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        category: { $ref: "#/components/schemas/UpdateCategory" }
        content: { type: string, minLength: 1, maxLength: 10000 }
        imageUrl: { type: ["string", "null"], format: uri }
        published: { type: boolean }
    UpdateUpdate:
      type: object
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        category: { $ref: "#/components/schemas/UpdateCategory" }
        content: { type: string, minLength: 1, maxLength: 10000 }
        imageUrl: { type: ["string", "null"], format: uri }
        published: { type: boolean }
    ContactMessageInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string, minLength: 2, maxLength: 160 }
        email: { type: string, format: email }
        subject: { type: string, minLength: 2, maxLength: 200 }
        message: { type: string, minLength: 5, maxLength: 5000 }
    ContactMessageReceipt:
      type: object
      required: [received]
      properties:
        received: { type: boolean }
...