# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  name: string;
  email: string;
  /**
     * @minLength 2
     * @maxLength 200
     */
  subject: string;
  /**
     * @minLength 5
     * @maxLength 5000
     */
  message: string;
}