# lib/api-zod/src/generated/types/updateCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
 * OpenAPI spec version: 0.1.0
 */

export type UpdateCategory = typeof UpdateCategory[keyof typeof UpdateCategory];


export const UpdateCategory = {
  Announcement: 'Announcement',
  News: 'News',
  Important_notice: 'Important notice',
  Program_update: 'Program update',
  Opportunity_update: 'Opportunity update',
  Event_reminder: 'Event reminder',
} as const;